삭제전 confirm추

+++ public/js/index.js
@@ -1,8 +1,10 @@
 function deleteajax(id) {
-    $.ajax({
-        url: '/country/' + id,
-        method: "DELETE",
-    }).done(function (response) {
-        location.href = "/";
-    });
+    if (confirm("삭제 하시겠습니까?")) {
+        $.ajax({
+            url: '/country/' + id,
+            method: "DELETE",
+        }).done(function (response) {
+            location.href = "/";
+        });
+    }
 }
